Allow topic search with partial name
Partners use a name convention to get the latest version of RHEL.
For example to get the latest nightly of RHEL-8 they use RHEL-8-nightly in the name.
Today RHEL-8-nightly is a real topic but will be remove in the future.

This patch introduce the .Y feature.
If a partner use RHEL-8.Y in the name, he will get the latest RHEL-8 topic.
To get the latest nightly he will use the filter tag param already in production.

The main difference is that RHEL-8.Y doesn't exists.
The control server replace the name with the latest RHEL-8 available.

Change-Id: If4b90921cdc631279f525adc3f8a3296e329b00a