Allow user creation with only sso_username
sso_username is the unique Red Hat login SSO given during the SSO authentication.
We want to simplify the onboarding process.
We are going to allow EPMs or Admin to create a user with only the Red Hat login.

This patch change the permission to EPM and admin instead of admin only.
It also change the email field in the user table to be empty.

Change-Id: Ibc7360594e540063b83ac5dfd04f4e269355d00d