Using gunicorn to run the API server
Since this is a wsgi compatible app, we can just hook it up to gunicorn.
I am building with a smaller container image and installing mostly from
source, this yields a much leaner result image. I am using the default
pip3 version because some of the libraries in requirements.txt (psycopg2
for example) don't work with newer versions of py3

Change-Id: I9aab626db0fdedd8c9a35282de7cb201a95de924