Add suject identifier for SSO users
We are going to change the format of the SSO token.
The new SSO token will contain information such as email_verified, and the sub (subject identifier) field which will be unique for each user.
This field will use the service user id instead of the unique keycloak id.
This field will replace sso_username and will be used to update user's information.

This patch adds the sso_sub field and populates it when we detect the new token.

This patch also modify the audiences used in the token decoding to be able to decode old tokens (`aud: dci`) and new tokens (`aud: ['api.dci', 'dci', ...]`) or future tokens (`aud: ['api.dci', ...]`)

Change-Id: I0aa8ac3c46778dc838795a632b711e32da5fd45d