Fix query_dsl for `not_contains` keyword
We received some queries which failed with an HTTP 500 with the
following error:

```
  File "/opt/dci-control-server/dci/db/query_dsl.py", line 109, in _build
    res = not m_column.contains([value])
  File "/usr/local/lib64/python3.6/site-packages/sqlalchemy/sql/elements.py", line 3090, in __bool__
    raise TypeError("Boolean value of this clause is not defined")
```

Queries looked like this (prettified)
```
/api/v1/jobs?query=
 and(
    contains(tags,daily)
    ge(created_at,2025-10-29)
    lt(created_at,2025-10-30)
    or(
      eq(remoteci_id,<some_remoteci_id>)
      and(
         eq(remoteci_id,<some_other_remoteci_id)
         not_contains(tags,<some_tag>)
      )
    )
)
&limit=200&offset=0&sort=-created_at
```

The code which translates `contains` and `not_contains` DSL keywords did
not have any tests and it appears that the translation for
`not_contains` was invalid.

Change `not column.contains()` to `sql.not_(column.contains())` does the
trick.

Change-Id: I79ba8986246211cb9ec2f50c962b5b868b476180